test(io_manager): 33 个集成测试 + 修复协程生命周期问题
新增:
- tests/test_iomanager.cpp: timer / fd / 并发 33 个用例
- CMakeLists.txt: test_iomanager 目标

修复:
- Task::destroySafe 区分 detached / fa 的情况,避免 UAF 和泄漏
- Scheduler::schedule(handle) 在协程完成后 destroy 句柄
- TimerManager::processExpired 消费句柄后清空,避免重复
- Scheduler::wait_all 用 wait_for 兜底,避免丢失唤醒
- ScopedlockedImpl::m_locked 初始化

验证:
- ASAN 无 UAF / leak
- UBSan 无 runtime error
- 33/33 测试通过

+++ blue/bytearray.cpp
@@ -656,7 +656,7 @@ namespace blue
 
     void ByteArray::addCapacity(size_t size)
     {
-        int remain = getCapacity();
+        size_t remain = getCapacity();
         if (size == 0 || remain >= size)
         {
             return;
